requirements.txt
numpy==2.1.3
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
websocket-client==1.8.0
xlwings==0.33.3
